Corrected profile delete option by adding associated feeder delete function too

+++ controllers/feederRoutes.js
@@ -136,12 +136,20 @@ router.post('/profile', async (req, res) => {
 // Router to delete the Pet based on Id. Passing the Id in the body and not as params.
 router.delete('/deletepet', async (req, res) => {
     try {
-        const deletePetDb = await Pet.destroy({ 
+        const deleteFeeder = await Feeder.destroy({
             where: {
                 pet_id: req.body.petId,
+                feed_date: new Date(),
             }
         });
-        res.status(200).json(deletePetDb);
+        if (deleteFeeder) {
+            const deletePet = await Pet.destroy({ 
+                where: {
+                    pet_id: req.body.petId,
+                }
+            });
+            res.status(200).json(deletePet);
+        }
     } catch (err) {
         res.status(500).json(err);
     }
